third_party/blink/web_tests/http/tests/loading/onload-vs-immediate-refresh-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
ALERT: SUCCESS
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - BeginNavigation request to 'http://127.0.0.1:8000/loading/resources/notify-success-on-load.html', http method GET
main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
You should have seen an alert.
